Fix static code analysis & wind speed label in settings

+++ lib/data/models/air_pressure_model.dart
@@ -3,6 +3,6 @@ import 'package:flutter_next_weather/domain/entities/air_pressure.dart';
 class AirPressureModel extends AirPressure {
   AirPressureModel.fromJson(Map<String, dynamic> json)
       : super(
-          json['air_pressure'],
+          json['air_pressure'] as double,
         );
 }
