+++ lib/data/models/wind_speed_model.dart
@@ -3,6 +3,6 @@ import 'package:flutter_next_weather/domain/entities/wind_speed.dart';
 class WindSpeedModel extends WindSpeed {
   WindSpeedModel.fromJson(Map<String, dynamic> json)
       : super(
-          json['wind_speed'],
+          json['wind_speed'] as double,
         );
 }
